refactor(schema): split postgres adapter modules
